Relicense 5 source files that are not part of the webapp from AGPL to GPL.
Building w/o the webapp is not supposed to pull in any AGPLed files.

I appear to have written all the code in these files;
the only commit by anyone else is 64e844e1fefca513d143e3f74a54dfc6d526a2ff
and is a spelling fix that is not copyrightable.